Add files via upload
This C++ project is a simple console-based program that simulates a hotel's inventory and sales management system. It allows users to order food items from the menu ,book a room keeps track of available quantities, and calculates the total sales for the day. The program offers a menu with options to order various food items and provides information about sales and remaining quantities.